requirements.txt
Pillow
imagehash
Send2Trash
